requirements.txt
pytest==7.4.3
pytest-allure-adaptor==1.0.7
allure-pytest==2.13.2
requests==2.31.0

